requirements.txt
plotly
streamlit_option_menu
openpyxl
